Generate two EC key pairs for LSKF Recovery
This CL is a basic version of generating key pairs for LSKF recovery
service. These two keys are generated on the same field and curve. The
only requirement they have is that the addition of their private keys
should not be divisible by the curve's prime number p. Currently this
code is not being called from anywhere in the codebase. AndroidKeyStore
has not been used as the key generator provider due to the fact that
AndroidKeyStore does not allow using the scalar version of the private
key for further calculations.

Bug: 454324424
Test: atest LskfResetTests
Flag: android.app.lskfreset.flags.enable_lskf_reset_manager

Change-Id: I2a761cd02f9f6fdd378e1991e04118d3eed2bc1d