Update HsumDefaultConfigOverlay config values
HsumDefaultConfigOverlay didn't touch the value of
config_enableMultiUserUI, which is false by default. So HSUM OEMs must
explicitly toggle it to true. This is silly, since in almost any
conceivable HSUM scenario, you'd want to allow multiple switchable
users. (If an HSUM OEM has some anomalous use-case for it, they can
always set it themselves.) While we're at it, we clarify what
config_enableMultiUserUI means, since the current description no longer
captures its definition.

Also removes config_isMainUserPermanentAdmin from
HsumDefaultConfigOverlay. This is already set to true by default, so it
doesn't matter. But HSUM devices may not wish to have a MainUser, and
setting explicitly here to true might therefore cause confusion and
conflict in inheritance. The goal is for HSUM to not require it, so we
shouldn't enable it by default here.

Fixes: 457528491
Fixes: 457524598
Fixes: 457540041
Test: N/A. Just javadoc and a "redundant" (at this point) configuration.
Flag: EXEMPT BUGFIX
Change-Id: I7458df122616f5e26d5e4e1040a79b7a3086b547