Fix potential NPE in AccessibilityEvent.init
This change prevents a potential NullPointerException in the `init` method
of `AccessibilityEvent` by adding a null check in the `appendRecord` method.

The `appendRecord` method did not previously check if the record being
added was null. This could lead to a `NullPointerException` in the `init`
method when it tried to create a new `AccessibilityRecord` from a null
entry in the `mRecords` list.

On HWASAN-enabled devices, this NullPointerException can manifest as a
more severe crash with a "tag-mismatch" error, as the runtime's attempt
to create and throw the NPE can access corrupted memory.

This change introduces a null check in `appendRecord` to prevent null
records from being added to the list, thus avoiding the potential crash.
The fix is gated by the feature flag `preventA11yEventNullRecord`.

Bug: 454269244
Test: atest AccessibilityEventTest
Flag: android.view.accessibility.prevent_a11y_event_null_record
Change-Id: Ia603274fa127528d64fbef60b14946a414a55901