[8/n] Integrate DisplayController in CompatUISharedStateHandler
DisplayController is integrated in the CompatUISharedStateHandler to
calculate stableBounds for the specific task. This will be used
by components to understand their position.

A areParentBoundsChanged property has been added to CompatUISharedState
to detect if the stableBounds are changed and the component will need to
update its position.

The Unified Testing Pattern has been implemented for the CompatUIHandler
abstraction and used for CompatUISharedStateHandler.

Flag: NONE Refactoring
Bug: 365691438
Test: atest WMShellUnitTests:CompatUISharedStateHandlerTest

Change-Id: Ia6c56c0b12fec70727d1870db3dffc26465a0e15