[DeveloperVerifierController] Shorten disconnect timeout
Shorten the verifier disconnect timeout to 30 seconds to allow for
faster resource cleanup. Add more debug logging to trace the
auto-disconnect behavior.

Also, when a connection fails, the status tracker is removed to ensure
that the timeout countdown is properly stopped without having to go
through the whole countdown.

Bug: 448139516
Test: manually tested and checked that the verifier is disconnected in time
FLAG: EXEMPT BUGFIX

Change-Id: Ie706f469bfd20ce01c4213d39d6b234f4b86e666