psc: Push hasExternalProcessHandles to ContentProviderRecordInternal
Previously, the `hasExternalProcessHandles` state was calculated
on-demand  when needed by the OomAdjuster. `ProcessStateController` had
methods to directly add/remove external provider clients, which in turn
called into `ContentProviderRecord`.

This CL refactors this to a push model. `ContentProviderRecord` is now
fully responsible for managing its external process handles. When the
number of handles changes, it notifies `ProcessStateController` with the
new boolean state of `hasExternalProcessHandles`.

This improves encapsulation by making `ContentProviderRecord` the single
source of truth for this state and simplifies the interface of
`ProcessStateController`.

Bug: 457589118
Test: m services.core
Test: atest MockingOomAdjusterTests ActivityManagerServiceTest
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: Ic96b8988b246b92ed8bea4efca7b903398340693