Replace NSSL drawbounds with stackbounds
DrawBounds was a concept for setting a drawable Rect on the NSSL, but this was difficult to model during HUN-to-Shade transitions. The horizontal bounds were also never used by the NSSL, so storing those values was unnecessary.

This CL replaces the DrawBounds Rect with a pair of top and bottom positions.

Bug: 457958356
Test: atest PlatformScenarioTests
Flag: com.android.systemui.scene_container

Change-Id: I7382c8468a7b6892cf853cf16d558c4888cd8db3