Reduce noise of test by WAKE transition on non-default display
The waitHandlerIdle(mH) is not enough because the transition is
ready from animation thread and then posts to display thread by
FallbackPlayer (if enabled).

It is not friendly that each tests need to be aware of there will
be an implicit transition that may queue some operations to cause
the test fail. So make most of test display usages not sleeping to
avoid triggering the WAKE transition when adding the first task.

Fix: 454240030
Flag: EXEMPT TEST_ONLY
Test: TaskTests#testFullscreenBoundsForcedOrientation
      AppCompatAspectRatioOverridesTest# \
      testSystemFullscreenOverride_activityRestartedInDefaultDisplay_true
      PresentationControllerTests

Change-Id: Ifc87aa2b75cd87869865d7661827e4ab07778f66