Add transition-based state machine for self-kill recovery
Currently, our heuristics for self-kill recovery are very basic in
that the activity relaunch caused by a configuration change caused
by a display move isn't tracked properly.

This change adds some transition-tracking logic with an internal
state machine so that we can more precisely identify the self-kill
caused by display move. This change consists of the following
sub changes:

[Transition tracking related to display move]
With our heuristics, we want to detect whether restart happens
during "display move". Tracking the display move transition fits
this purpose. Active display-move transitions are maintained in
mDisplayMoveTransitions, and we clean this up by registering the
end listener (also we schedule our own timeout callback).

[Self-kill state machine]
Based on mDisplayMoveTransitions, we can create a simple state
machine to check if the activity is finishing due to an activity
relaunch caused by a config change caused by display move.
The machine state transitions from UNDEFINED -> DISPLAY_MOVING
-> RELAUNCHING_ON_DISPLAY_MOVE -> SELF_KILLING_ON_RELAUNCH to
track this whole process.

[Config change check]
To ensure we won't apply our self kill recovery when not needed,
this change also introduces a check if the activity restart is
caused by a combination of display-related config changes.

[Changes in where to hook in activity code]
Previously we hooked onActivityFinishing() in
makeFinishingLocked(), but this only covers the case of
Activity.finish(), but not Activity.finishAndRemoveTask().
This change moves the hook to
ActivityClientController.finishActivity().
Also, this change removes the onActivityDestroyed() hook, as
this is no longer need now that we listen to shell transition
end events, which wait for the activity to be resumed back or
killed.

Flag: com.android.window.flags.enable_auto_recovery_from_self_kill
Bug: 446998828
Test: AppCompatDisplayCompatTests
Change-Id: I76914325d7ba51c41affab1a1e432128cc9f2ab6