Add preinstall allowlisting for local transport and backup confirmation
BackupRestoreConfirmation is the UI that pops up when running `adb
backup` and `adb restore` to let the user confirm a local backup or
restore operation should be performed. It can be used by full users and
in the work profile.

The LocalTransport is a BackupTransport intended to be used for testing.
When selected as the active transport, the `BackupManagerService` will
back up and restore to/from the local disk. The `BackupManagerService`
creates a `UserBackupManagerService` for every full user on the device
and also includes the work profile.

Bug: 450268260
Bug: 450268637
Test: verified `adb backup` works for all users on HSUM device
Test: verified the localtransport can be selected for all users on HSUM device
Test: adb shell dumpsys package com.android.sharedstoragebackup | grep "installed=true"
Flag: NONE allowlist change only

Change-Id: I025005402ae993fbcb2f6dd3fcd79d3e5ecb0659