Simplify switching users in IMMS
Currently in IMMS we first verify if there is any pending user switch.
If there is, and the switch is for the given user, we call schedule
again just to update the mClientToBeReset. If the switch is for a
different user, we verify if the given user is a profile of the
mCurrentImeUserId. This was introduced in [1]. However, the list of
profiles of an userId will also contain the userId itself.

If we have a pending switch from user 0 to user 10, and we get another
startInput call on user 0 before the switch is processed, we will cancel
it, and schedule a new switch from user 0 to user 0, which is wasteful.

This simplifies the user switching by only checking if the given user is
already the mCurrentImeUserId, and if so cancels the pending switch. If
this is a different user, we should fallback to actually checking
whether it is a profile of the current user, and then schedule a new
switch, regardless of having a pending switch or not. This will also
have the benefit of logging startInput call on users that are not
profiles of the current user, and sending an IME hide request on their
clients, even if we have a pending user switch.

  [1]: I5a73a66d2b8acadad9b3577ebc4c17b5a25fd011

Flag: EXEMPT BUGFIX
Test: CtsInputMethodTestCases MultiUserMockImeTests
Bug: 281029564
Change-Id: I09df862f15f9bef4221acd27b02342265ce109e3