Do not skip default desk creation for DEFAULT_DISPLAY on keyguard unlock
This was originally added in #onKeyguardVisibilityChanged becase
potential reconnect is not needed for the default display, but it also
had the effect of skipping default desk creation if needed.

Instead, move the display check into #handlePotentionalReconnect, so
that reconnection never happens for the default display AND the
reconnection failing makes the display one of the candidates for default
desk creation in #onKeyguardVisibilityChanged.

Flag: EXEMPT BUGFIX
Fix: 445996567
Test: atest DesktopDisplayEventHandlerTest
Change-Id: Iccc4e73e9d0db2d405531b60af7fdd3b61ccc54e