Honor dream wireless charging restriction config
PowerManagerService was not respecting the `config_supportDreamWirelessChargingRestriction` configuration. This meant that the setting to restrict dreaming to wireless charging was being honored even on devices where wireless charging was not supported.

This change updates PowerManagerService to check the value of `config_supportDreamWirelessChargingRestriction`. If the config is false, the setting is ignored, and the restriction does not apply. This mirrors the existing logic in DreamManagerService.

Unit tests have been added to PowerManagerServiceTest to verify this new behavior.

Bug: 457479588
Test: atest PowerManagerServiceTest
Flag: EXEMPT BUGFIX
Change-Id: Ia146eea9d338e453aff4681afd934ec031b59e82