Enhance the assertion and comments of `testWeOnlyShrinkIfCapacityWillChange`.
The test enhanced is ensuring that the heap does not perform a wasteful array reallocation when the calculated shrink size is identical to the current capacity. This CL enhances the comments of it. It also adds an assertion of the heap capacity - the capacity should not be changed if the heap does not shrink.



This CL is generated by AI, please check the comments for more details.

Test: ATP tests passed http://go/forrest-run/L60800030035787485
Bug: 431235865 (Umbrella)
Flag: TEST_ONLY



Change-Id: I63f6ca2412d33137d053d45c603780b5f7d01e0b