Separate the wait condition for notification and clickable
Second attempt at trying to fix the missing notification issue.
The test is still failing with notification not found.
Separate the check for notification and whether it is clickable.

Hopefully this gives more insight into whether notification is not found or that it is not clickable.

Bug: 452985189
Flag: EXEMPT TEST_ONLY
Test: monitoring flake
Change-Id: I1019babb4e554e73cac8b050c182a9eb3016baf3