Policy Streamlining: Support 'getPolicy' in generic handler of screen capture policy
I also changed the behavior of 'setPolicy' for screen capture disabled.
Previously it followed how things were set in the manually written handler,
but it makes more sense to me that it follows how we want things to be
(as long as that doesn't fundamentally conflicts with the old behavior, which it doesn't).

Bug: 445147116
Test: atest FrameworksServicesTests:com.android.server.devicepolicy.handlers.PolicyHandlerTest
Flag: android.app.admin.flags.policy_streamlining
Change-Id: I18541e0aa0c4963098cc67cbd0d4d00f163cd632