Adjust PIN bouncer in BelowUserSwitcherLayout
Adjust the BelowUserSwitcherLayout in compose, making it more adaptable
to different display sizes. This is particularly important in EXPANDED
heights that are not so large (e.g. 900px). As described in the
associated bug, the PIN buttons didn't have enough space and looked
squeezed in such screen sizes.

This CL significantly reduces the sizes of the fixed vertical padding.
They are replaced with Spacers, that dynamically expand to fill in the
available space.

Test: manual - verified bouncer UI on multiple display sizes that trigger the BELOW_USER_SWITCHER layout
Bug: 418961289
Flag: EXEMPT BUGFIX
Change-Id: If37abbcacc673f516e331b54a42c8988dba5c1a9