Enable ContextTest#testIsUiContext_appContext_returnsFalse on Ravenwood
This test was previously disabled on Ravenwood because the underlying logic for determining if a context is a UI context was not implemented.

With the necessary dependencies now available in the Ravenwood test environment, this test is enabled to validate that `isUiContext()` correctly returns false for a standard application context. This improves test coverage for this common use case on Ravenwood.



This CL is generated by AI, please check the comments for more details.

Test: ATP tests passed http://go/forrest-run/L48200030036853703
Bug: 431235865 (Umbrella)
Flag: TEST_ONLY



Change-Id: I028fffbe7ecace4e71d8e9dda995060fb4c3d504