psc: Move SDK sandbox connection logic to base class
This CL refactors the management of SDK sandbox connections by moving the
relevant logic from `ProcessServiceRecord` to its abstract base class,
`ProcessServiceRecordInternal`.

Previously, the `mSdkSandboxConnections` collection and its accessor
methods were implemented solely in the concrete `ProcessServiceRecord`
class. This change moves this state and functionality into the shared
`ProcessServiceRecordInternal` class, centralizing the logic.

As a result, `ProcessServiceRecord` is simplified, and `OomAdjusterImpl`
is updated to call the new method on the base class.

Bug: 457589118
Test: m services.core
Test: atest MockingOomAdjusterTests ActivityManagerServiceTest
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I6e3440aaef97bb0acaf6e48fa82daa7743a4f296