psc: Move host process field to ServiceRecordInternal
The `app` field, which represents the host process for a service, was
previously a public field in the `ServiceRecord` class. This required
components that operate on the `ServiceRecordInternal` abstraction, such
as the OomAdjuster, to depend on the concrete implementation to access
the host process.

This CL refactors this by pulling up the host process field into the
`ServiceRecordInternal` base class. The field is made private and
accessed via `getHostProcessInternal()` and `setHostProcess()`.

Bug: 457589118
Test: m services.core
Test: atest MockingOomAdjusterTests ActivityManagerServiceTest
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I5097457bad2d3b5199b68e493f9793a81f71c8f3