Move VolumeNavigator's initialization out of its construction.
Without this, VolumeNavigator was not properly initalizded unless
someone happened to construct it, which happened coincidentally in
VolumeModule.

Had the old VolumeDialogImpl been deleted, and VolumeModule updated to
reflect this, then VolumeNavigator would not have been constructed,
and we would have seen breaking behavior with the volume.

This actually happened when http://ag/36782089 was submitted, resulting
in a P0 rollback.

Bug: 219008720
Flag: EXEMPT PURE_REFACTOR
Test: manual
Change-Id: I295515cb8b3b2d156237c5e1fa25f292551f7678