Only skip snapshot during bounds change to different display
Previously, a screenshot would not be taken if the bounds change to
prevent a screenshot from being taken during a transition from visible
to invisible. This happens when we have a bounds change where the task
is moving displays. The issue with this is that screenshots were not
taken during every biunds change transition which resulted in blank
tiles in overview due to missing screenshots.

This change only prevents the screenshot when there is a bounds change
and the task is moving to a different display.

Bug: 452456436
Test: manual testing
Flag: EXEMPT BUGFIX
Change-Id: I85ef7b91b9cea59852b681035b0262d0fe73fe65