Decommission sqlite micro benchmarks
SQLite benchmarks are very sensitive to AFDO profiles and regress
whenever the profiles are invalidated, which occurs whenever there is
any change to the external/sqlite code.  The regression always clears
when the profiles are updated, but this can take weeks.  No actionable
regressions have been detected by the microbenchmarks in the last five
years; decommissioning them now.

The full list of decommissioned tests is:
 * testSelect
 * testSelectCacheMissRate
 * testSelectMultipleRows
 * testCursorIterateForward
 * testCursorIterateBackwards
 * testInnerJoin
 * testInsert
 * testInsertWithPersistFull
 * testInsertWithWalNormalMode
 * testInsertWithWalFullMode
 * testBulkInsertWithWalNormalMode
 * testBulkInsertWithPersistFull
 * testBulkInsert
 * testDelete
 * testUpdateWithWalNormalMode
 * testUpdateWithPersistFull
 * testUpdate
 * testMultithreadedReadWrite
 * testMultithreadedReadWriteWithWalNormal

Additional tests may be decommissioned the next time regressions are
reported because AFDO profiles are out of date.

Flag: EXEMPT TEST_ONLY
Test: presubmit
Bug: 457836285
Change-Id: I1a60d56500019360107281bc9faa81d3de31a82b