Add hidden drag flag to disable default pointer icon.
This CL adds a hidden flag to disable the default pointer icon that is
normally set on drag start when the input source is a mouse. This will
be used by Launcher to normalize the pointer icon used when dragging
workspace items since the Launcher supports both internal-level and
system-level drag-and-drop and doesn't wish to use the default pointer
icon.

Bug: 456490978
Flag: EXEMPT BUGFIX
Test: atest DragDropControllerTests
Change-Id: I00a819fe0fb27dd6b4c38d825c61db4e41977834