[flexiglass] PIN bouncer pad button touch
Note: This reverts commit aa53e59831401ed54130467bc3b309f8e3246c39 or a redo of the original, ag/36775663. The CTS test breakage that caused it to be reverted will be fixed in a follow-up, as well as b/458604577 - which was caused by the original CL.

---

Solves the problem of allowing faster, less careful clicks without
creating a drag gesture that would dismiss the bouncer.

Fix: 454368363
Test: manually tested. Made sure that tapping fast works, long presses
work, accessibility, haptics, and indication animations all work
Flag: com.android.systemui.scene_container

Change-Id: If110eff1b83f53069ff607d9c04d59f5e2ef72d5