Only show desktop mode option and its settings when desktop is supported
In SystemUI: when connecting a display, only show the desktop/extend
option in the connection preference dialog when desktop mode is
supported.

In DisplayManager:
1 - On startup, override the mirror/extend setting to be mirror if
desktop mode is not supported.
2 - When returning the connection preference, check if desktop mode is
supported, and override it if it was desktop/extend, and desktop mode is
not supported.

In Settings: only show mirror/extend related settings if desktop mode is
supported.

Bug: 450957248
Flag: EXEMPT Bugfix
Test: atest DisplayServiceTests:DisplayManagerServiceTest
Test: atest SystemUITests:ExternalDisplayConnectionDialogDelegateTest
Test: Disable desktop mode config overlay and try plugging in a
      display. Then check Settings as well
Change-Id: I7048360e290df813945313cb0c2b0555e2c9c4cb