[PiP on DW] Fix drag-PiP-across-display crash for Chrome.
There is a Chrome-specific issue on PiP CD where dragging a Chrome PiP
to a new display will trigger PiP removal via a TO_BACK transition
requested by the Chrome app. This new transition causes a crash due to
the EXITED_PIP PipTransitionState causing the PiP leash to be set to
null. When the drag-PiP-across-display animation finishes, it tries to
access the leash and gets a NPE.

This CL adds a special handling to cancel the ongoing animation and
finish the drag-PiP-across-display transition if we get a merging
transition that removes the PiP.

Note that dragging a Chrome PiP to a new display will still lead to the
PiP removal, but this is WAI since the app is explicitly requesting it.

Bug: 458167023
Test: atest PipDisplayTransferHandlerTest
Test: Manual - enter Pip on Chrome and drag to external display. Verify
that there is no crash and PiP is removed correctly.
Flag: EXEMPT BUGFIX

Change-Id: I91cac9e389c351795527101fc4647fe26efbe648