Prevent TaskLPM from overwriting desktop-first policy
This CL prevents TaskLPM from overwriting the fullscreen decision of
desktop-first policy.
If the desktop-first policy decides the windowing mode to FULLSCREEN
(instead of UNDEFINED), it has an intention so we should keep it as a
final result.

Flag: com.android.window.flags.enable_desktop_first_policy_fullscreen_decision_bugfix
Bug: 383617953
Test: CtsWindowManagerDeviceInput
Change-Id: I4d41041ccd4762d62a3260821d9e1d01f7f9d8a3