TrustedPresentationListenerController: exclude ComputerControl displays
Content on ComputerControl displays might not be presented to the user except through ComputerControl's MirrorView. If the MirrorView is presented, it attaches a cloned hierarchy to the display it's presented on, and the TPL machinery will work through that - except that the virtual display's instance will step on that state and so it never stabilizes.

As a short-term fix, we'll exclude ComputerControl displays for TPL purposes, so that windows are only considered presented if they are presented through the MirrorView. Longer term, we need to address how TPL and virtual displays interact.

Fixes: 460370389
Bug: 459548175
Bug: 459545101
Test: presubmit
Flag: android.companion.virtualdevice.flags.computer_control_access
Change-Id: I8d0c25f56b4fadd2d2e5148d2fc4c758905e3b4e