Check running animation for releasing transaction
When there are 2 running animations, if one of them is finished
first, the end callback may be called before scheduling the other
running animations. That causes to release the transaction that
will be used later. So a counter is added to avoid the case.

Bug: 456665218
Flag: com.android.window.flags.default_animator_single_transaction2
Test: DefaultTransitionHandlerTest
Change-Id: I380eb7ebfc5a2a82993240a184c6765e55b4a5e2