Add AiSeal feature to PackageManager
Add android.software.aiseal feature to PackageManager to be able to
distribute applications based on AiSeal feature availability on device.

Applications need to be implemented and compiled specifically to work
inside the AiSeal virtual machine. Without this feature, the application
will not be able to work.

Bug: 455594108
Test: TH, ABTD
Flag: android.aiseal.aiseal_host_apis
Change-Id: I59a60ddc8ee2bb65aac5a39348755a71b5ac0c5f